Add files via upload
This project is a machine learning-based Disease Prediction System designed to classify and predict diseases based on user-provided symptoms. It combines the precision of Support Vector Machines (SVM) with the interpretability of LIME and SHAP, the robustness of Jaccard Similarity, and the adaptability of Fuzzy Logic to handle uncertainty in symptom reporting.

The model achieves an impressive 99% accuracy, making it a reliable system for clinical decision support and academic research.